refactor: 와인 패키징 수정
