+++ src/main/java/com/drinkeg/drinkeg/domain/wine/repository/WineRepositoryImpl.java
@@ -3,7 +3,6 @@ package com.drinkeg.drinkeg.domain.wine.repository;
 import com.drinkeg.drinkeg.domain.wine.domain.Wine;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.types.dsl.Expressions;
-import com.querydsl.core.types.dsl.StringExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
